backend/requirements.txt
fastapi
uvicorn
requests
pandas
yfinance
python-dotenv
yagmail
schedule
openai
